camera_module/requirements.txt
absl-py==0.9.0
astor==0.8.1
bcolz==1.2.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
dnspython==1.16.0
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
Pillow==6.2.2
pkg-resources==0.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.10.1
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==2.0.2
tensorflow==2.0.1
tensorflow-estimator==2.0.1
tensorflow-gpu==2.1.0
tensorflow-metadata==0.21.1
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
